// src/app/entities/activity/models/activity.model.ts
export interface Activity {
  url?: string;
  id?: number;
  products?: any[];
  contact?: any;
  site_uuid: any[];
  actual_start_date?: any;
  actual_end_date?: any;
  actual_duration?: any;
  actual_cost?: any;
  address?: any;
  capacity_built?: any;
  description?: any;
  description_of_community_involvement?: any;
  description_of_government_involvement?: any;
  expected_end_date?: any;
  expected_start_date?: any;
  issues_and_challenges: any;
  justification_background?: any;
  lessons_learned?: any;
  level2_uuid?: string;
  name: string;
  notes?: any;
  parent_workflowlevel2?: any;
  quality_assured?: any;
  risks_assumptions?: any;
  short_name?: any;
  site_instructions?: any;
  total_cost?: any;
  total_estimated_budget?: any;
  type?: any;
  effect_or_impact?: any;
  create_date?: string;
  edit_date?: string;
  status?: any;
  progress?: any;
  donor_currency?: any;
  local_currency?: any;
  milestone?: any;
  office?: any;
  sector?: any;
  staff_responsible?: any;
  workflowlevel1: any;
  created_by?: any;
  approval?: any[];
  indicators?: any[];
  site?: any[];
  stakeholder?: any[];
  sub_sector?: any[];
}